Major refactor, see commit body
I was not very happy with how the guilds were saved in the DB. It meant
that if there was 2 channel in one guild that where bound, we would keep
trap of 2 copy of the guild.

Now, channel and it's subjects are bound to an array of combinations
that are held in one guild within the database. This allows to have
different subjects in different channels.

Also, the in-memory article store was changed to use the DB in order to limit
memory usage as the bot runs for months at a time.