Skip articles without a direct link
Sometimes, an article on Hackernews can be a discussion or a hiring
post. In that case, the article's link is empty, which means that we
can't index it properly to avoid sending it repeatedly.

This fix simply skips any post that does not contains a link to an
outside source. We simply ignore hiring, discussion, etc. As those are
not really "Tech News" worthy anyway.